fix: created video player that is a drop in replacement for react player with no internet error
