+++ packages/cysync-core/src/dialogs/Inheritance/SilverPlanPurchase/Dialogs/Instructions.tsx
@@ -6,7 +6,7 @@ import { selectLanguage, useAppSelector } from '~/store';
 
 import { useInheritanceSilverPlanPurchaseDialog } from '../context';
 import { Layout } from '../Layout';
-import { CustomReactPlayer } from '~/components/CustomReactPlayer';
+import { VideoPlayer } from '~/components/VideoPlayer';
 
 export const Instructions = () => {
   const lang = useAppSelector(selectLanguage);
@@ -36,8 +36,10 @@ export const Instructions = () => {
       >
         <LangDisplay text={strings.instructions.video.title} />
       </Typography>
-      <CustomReactPlayer
+      <VideoPlayer
         url={constants.inheritance.silverPlanPurchaseTutorialLink}
+        width="720px"
+        height="405px"
       />
     </Layout>
   );
